feat(api): Rename accreditation-repository to complementary-certification-repository
