+++ api/lib/domain/usecases/index.js
@@ -1,6 +1,5 @@
 const dependencies = {
   accountRecoveryDemandRepository: require('../../infrastructure/repositories/account-recovery-demand-repository'),
-  accreditationRepository: require('../../infrastructure/repositories/accreditation-repository'),
   answerRepository: require('../../infrastructure/repositories/answer-repository'),
   assessmentRepository: require('../../infrastructure/repositories/assessment-repository'),
   assessmentResultRepository: require('../../infrastructure/repositories/assessment-result-repository'),
@@ -40,7 +39,6 @@ const dependencies = {
   certificationCourseRepository: require('../../infrastructure/repositories/certification-course-repository'),
   certificationCpfCityRepository: require('../../infrastructure/repositories/certification-cpf-city-repository'),
   certificationCpfCountryRepository: require('../../infrastructure/repositories/certification-cpf-country-repository'),
-  countryRepository: require('../../infrastructure/repositories/country-repository'),
   certificationIssueReportRepository: require('../../infrastructure/repositories/certification-issue-report-repository'),
   certificationLsRepository: require('../../infrastructure/repositories/certification-livret-scolaire-repository'),
   certificationOfficerRepository: require('../../infrastructure/repositories/certification-officer-repository'),
@@ -56,7 +54,9 @@ const dependencies = {
   competenceMarkRepository: require('../../infrastructure/repositories/competence-mark-repository'),
   competenceRepository: require('../../infrastructure/repositories/competence-repository'),
   competenceTreeRepository: require('../../infrastructure/repositories/competence-tree-repository'),
+  complementaryCertificationRepository: require('../../infrastructure/repositories/complementary-certification-repository'),
   correctionRepository: require('../../infrastructure/repositories/correction-repository'),
+  countryRepository: require('../../infrastructure/repositories/country-repository'),
   courseRepository: require('../../infrastructure/repositories/course-repository'),
   divisionRepository: require('../../infrastructure/repositories/division-repository'),
   encryptionService: require('../../domain/services/encryption-service'),
